Add `ignore` keyword argument to `copytree`
The `ignore` argument expects a callable that returns a list of
names that will be ignored while copying. It supports
`shutil.ignore_patterns` [1] to allow using a similar interface.

Also, include proper typing according to the behaviour described in
the Python docs on `shutil.copytree` and expanding it to support the
`cloudpathlib` environment.

[1]: https://docs.python.org/3/library/shutil.html#shutil.ignore_patterns

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>